+++ InvoiceDiskLast/Models/ExpenseDetailModel.cs
@@ -10,10 +10,10 @@ namespace InvoiceDiskLast.Models
         public int Id { get; set; }
         public int  EXPENSE_ACCOUNT_ID { get; set; }
         public string DESCRIPTION { get; set; }
-        public decimal  AMOUNT { get; set; }
-        public decimal  TAX_PERCENT { get; set; }
-        public decimal  TAX_AMOUNT { get; set; }
-        public decimal  SUBTOTAL { get; set; }
+        public double AMOUNT { get; set; }
+        public double  TAX_PERCENT { get; set; }
+        public double  TAX_AMOUNT { get; set; }
+        public double  SUBTOTAL { get; set; }
         public int expense_id { get; set; }
         public int user_id { get; set; }
         public int comapny_id { get; set; }
